Valgrind for mem usage of pytorch
This commit modifies the pytorch program and the
python evaluate script in order to collect memory
profiling information with the use of valgrind.

This helps us understand how much memory the pytorch
process uses and particularly, observe whether the
additional memory used during model loading is returned
to the OS.

Just run the evaluate script and valgrind will create
the massif report.